Implement expense saving
Maven no longer builds the Docker image of the project, we plan to use a
Dockerfile instead.

Expense saving to the database has been implemented. The method is
covered by tests using an experimental approach with nested classes for
now. Perhaps in the future, we will refactor it to make the tests
simpler and more readable.

Some DTOs and classes for handling data returned as pages have been
added.